Divided up Data.Functor.Linear
