+++ linear-base.cabal
@@ -37,7 +37,8 @@ library
     Data.Either.Linear
     Data.Eq.Linear
     Data.Functor.Linear
-    Data.Functor.Linear.Internal
+    Data.Functor.Internal.Functor
+    Data.Functor.Internal.Applicative
     Data.HashMap.Mutable.Linear
     Data.List.Linear
     Data.Maybe.Linear
