+++ src/Data/Functor/Internal/Functor.hs
@@ -1,21 +1,29 @@
 {-# OPTIONS_HADDOCK hide #-}
 {-# LANGUAGE LinearTypes #-}
 {-# LANGUAGE NoImplicitPrelude #-}
-
-module Data.Functor.Linear.Internal where
+module Data.Functor.Internal.Functor
+  (
+    Functor(..)
+  , (<$>)
+  , (<$)
+  , void
+  ) where
 
 import Prelude.Linear.Internal
 import Prelude (Maybe(..), Either(..))
 import Data.Functor.Const
 import Data.Functor.Sum
 import Data.Functor.Compose
-import Data.Monoid.Linear hiding (Sum)
 import Data.Functor.Identity
 import qualified Control.Monad.Trans.Reader as NonLinear
 import qualified Control.Monad.Trans.Cont as NonLinear
 import qualified Control.Monad.Trans.Maybe as NonLinear
 import qualified Control.Monad.Trans.Except as NonLinear
 import qualified Control.Monad.Trans.State.Strict as Strict
+import Data.Unrestricted.Internal.Consumable
+
+-- # Functor Defintion
+-------------------------------------------------------------------------------
 
 -- | Linear Data Functors should be thought of as containers holding values of
 -- type @a@ over which you are able to apply a linear function of type @a %1->
@@ -27,56 +35,24 @@ class Functor f where
 (<$>) :: Functor f => (a %1-> b) -> f a %1-> f b
 (<$>) = fmap
 
--- | Data 'Applicative'-s can be seen as containers which can be zipped
--- together. A prime example of data 'Applicative' are vectors of known length
--- ('ZipList's would be, if it were not for the fact that zipping them together
--- drops values, which we are not allowed to do in a linear container).
---
--- In fact, an applicative functor is precisely a functor equipped with (pure
--- and) @liftA2 :: (a %1-> b %1-> c) -> f a %1-> f b %1-> f c@. In the case where
--- @f = []@, the signature of 'liftA2' would specialise to that of 'zipWith'.
---
--- Intuitively, the type of 'liftA2' means that 'Applicative's can be seen as
--- containers whose "number" of elements is known at compile-time. This
--- includes vectors of known length but excludes 'Maybe', since this may
--- contain either zero or one value.  Similarly, @((->) r)@ forms a Data
--- 'Applicative', since this is a (possibly infinitary) container indexed by
--- @r@, while lists do not, since they may contain any number of elements.
---
--- == Remarks for the mathematically inclined
---
--- An 'Applicative' is, as in the restricted case, a lax monoidal endofunctor of
--- the category of linear types. That is, it is equipped with
---
--- * a (linear) function @() %1-> f ()@
--- * a (linear) natural transformation @(f a, f b) %1-> f (a, b)@
---
--- It is a simple exercise to verify that these are equivalent to the definition
--- of 'Applicative'. Hence that the choice of linearity of the various arrow is
--- indeed natural.
-class Functor f => Applicative f where
-  {-# MINIMAL pure, (liftA2 | (<*>)) #-}
-  pure :: a -> f a
-  (<*>) :: f (a %1-> b) %1-> f a %1-> f b
-  f <*> x = liftA2 ($) f x
-  liftA2 :: (a %1-> b %1-> c) -> f a %1-> f b %1-> f c
-  liftA2 f x y = f <$> x <*> y
+-- | Replace all occurances of @b@ with the given @a@
+-- and consume the functor @f b@.
+(<$) :: (Functor f, Consumable b) => a -> f b %1-> f a
+a <$ fb = fmap (`lseq` a) fb
 
----------------
--- Instances --
----------------
+-- | Discard a consumable value stored in a data functor.
+void :: (Functor f, Consumable a) => f a %1-> f ()
+void = fmap consume
 
--- Standard instances
+-- # Instances
+-------------------------------------------------------------------------------
+
 instance Functor [] where
   fmap _f [] = []
   fmap f (a:as) = f a : fmap f as
 
 instance Functor (Const x) where
   fmap _ (Const x) = Const x
-
-instance Monoid x => Applicative (Const x) where
-  pure _ = Const mempty
-  Const x <*> Const y = Const (x <> y)
 
 instance Functor Maybe where
   fmap _ Nothing = Nothing
@@ -89,16 +65,8 @@ instance Functor (Either e) where
 instance Functor ((,) a) where
   fmap f (x,y) = (x, f y)
 
-instance Monoid a => Applicative ((,) a) where
-  pure x = (mempty, x)
-  (u,f) <*> (v,x) = (u <> v, f x)
-
 instance Functor Identity where
   fmap f (Identity x) = Identity (f x)
-
-instance Applicative Identity where
-  pure = Identity
-  Identity f <*> Identity x = Identity (f x)
 
 instance (Functor f, Functor g) => Functor (Sum f g) where
   fmap f (InL fa) = InL (fmap f fa)
@@ -107,21 +75,12 @@ instance (Functor f, Functor g) => Functor (Sum f g) where
 instance (Functor f, Functor g) => Functor (Compose f g) where
   fmap f (Compose x) = Compose (fmap (fmap f) x)
 
-instance (Applicative f, Applicative g) => Applicative (Compose f g) where
-   pure x = Compose (pure (pure x))
-   (Compose f) <*> (Compose x) = Compose (liftA2 (<*>) f x)
-   liftA2 f (Compose x) (Compose y) = Compose (liftA2 (liftA2 f) x y)
-
 ---------------------------------
 -- Monad transformer instances --
 ---------------------------------
 
 instance Functor m => Functor (NonLinear.ReaderT r m) where
   fmap f (NonLinear.ReaderT g) = NonLinear.ReaderT (\r -> fmap f (g r))
-
-instance Applicative m => Applicative (NonLinear.ReaderT r m) where
-  pure x = NonLinear.ReaderT (\_ -> pure x)
-  NonLinear.ReaderT f <*> NonLinear.ReaderT x = NonLinear.ReaderT (\r -> f r <*> x r)
 
 -- The below transformers are all Data.Functors and all fail to be
 -- Data.Applicatives without further restriction. In every case however,
@@ -143,3 +102,4 @@ instance Functor (NonLinear.ContT r m) where
 
 instance Functor m => Functor (Strict.StateT s m) where
   fmap f (Strict.StateT x) = Strict.StateT (\s -> fmap (\(a, s') -> (f a, s')) (x s))
+
