+++ src/Data/Unrestricted/Internal/Consumable.hs
@@ -5,13 +5,10 @@ module Data.Unrestricted.Internal.Consumable
   (
   -- * Consumable
     Consumable(..)
-  , void
   , lseq
   , seqUnit
   )
   where
-
-import qualified Data.Functor.Linear.Internal as Data
 
 class Consumable a where
   consume :: a %1-> ()
@@ -26,8 +23,4 @@ seqUnit () b = b
 -- This is like 'seq' but the first argument is restricted to be 'Consumable'.
 lseq :: Consumable a => a %1-> b %1-> b
 lseq a b = seqUnit (consume a) b
-
--- | Discard a consumable value stored in a data functor.
-void :: (Data.Functor f, Consumable a) => f a %1-> f ()
-void = Data.fmap consume
