+++ src/Data/Unrestricted/Internal/Instances.hs
@@ -25,7 +25,8 @@ import Data.Unrestricted.Internal.Consumable
 import Data.Unrestricted.Internal.Dupable
 import Data.Unrestricted.Internal.Movable
 import Data.Unrestricted.Internal.Ur
-import qualified Data.Functor.Linear.Internal as Data
+import qualified Data.Functor.Internal.Functor as Data
+import qualified Data.Functor.Internal.Applicative as Data
 import GHC.Types hiding (Any)
 import Data.Monoid.Linear
 import Data.List.NonEmpty
