+++ src/Data/Unrestricted/Linear.hs
@@ -66,7 +66,6 @@ module Data.Unrestricted.Linear
   , Consumable(..)
   , Dupable(..)
   , Movable(..)
-  , void
   , lseq
   , dup
   , dup3
