+++ src/Data/V/Linear.hs
@@ -59,7 +59,8 @@ module Data.V.Linear
   , caseNat
   ) where
 
-import qualified Data.Functor.Linear.Internal as Data
+import qualified Data.Functor.Internal.Functor as Data
+import qualified Data.Functor.Internal.Applicative as Data
 import qualified Data.Traversable.Linear as Data
 import Data.Kind (Type)
 import Data.Proxy
